requirements.txt
flwr
torch
tenseal
streamlit
plotly
numpy
scipy
scikit-learn
pyyaml
pyserial
matplotlib